+++ src/ispypsa/cli/dodo.py
@@ -17,15 +17,15 @@ from ispypsa.data_fetch import (
 )
 from ispypsa.iasr_table_caching import build_local_cache, list_cache_files
 from ispypsa.logging import configure_logging
-from ispypsa.model import (
-    build_pypsa_network,
-    save_pypsa_network,
-    update_network_timeseries,
-)
 from ispypsa.plotting import (
     create_plot_suite,
     generate_results_website,
     save_plots,
+)
+from ispypsa.pypsa_build import (
+    build_pypsa_network,
+    save_pypsa_network,
+    update_network_timeseries,
 )
 from ispypsa.results import (
     extract_regions_and_zones_mapping,
