--- src/ispypsa/model/__init__.py
@@ -1,5 +1,0 @@
-from ispypsa.model.build import build_pypsa_network
-from ispypsa.model.save import save_pypsa_network
-from ispypsa.model.update import update_network_timeseries
-
-__all__ = ["build_pypsa_network", "save_pypsa_network", "update_network_timeseries"]
