+++ tests/test_model/test_initialise.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 import pandas as pd
 
-from ispypsa.model.initialise import _initialise_network
+from ispypsa.pypsa_build.initialise import _initialise_network
 from ispypsa.translator.snapshots import (
     _add_investment_periods,
     _create_complete_snapshots_index,
